refactor(settings): split Settings into 7 domain sub-objects (PR1 + Tier 1 of PR2)
Decomposes the 1,170-line Settings class into composition-based domain
classes to reduce recompile blast radius (issue #743). Each sub-object
is a standalone QObject with its own QSettings instance and Q_PROPERTYs;
Settings owns them and exposes them via Q_PROPERTY for QML access
(Settings.mqtt.mqttEnabled, Settings.theme.activeShader, etc.).

Domain classes added:
- SettingsMqtt        — MQTT broker/publish (10 properties)
- SettingsAutoWake    — wake schedule (4 properties)
- SettingsHardware    — heater calibration sent to firmware (6 properties)
- SettingsAI          — AI provider/keys (8 properties + configurationChanged signal)
- SettingsTheme       — theme/shader/font/brightness (~25 properties + invokables)
- SettingsVisualizer  — visualizer.coffee upload (8 properties)
- SettingsMcp         — MCP server (4 properties)

Also extracted shot history struct types (ShotRecord, HistoryShotSummary,
ShotFilter, ShotSaveData, GrinderContext, HistoryPhaseMarker) into
src/history/shothistory_types.h, allowing shotfileparser.h consumers
to skip the full shothistorystorage.h.

Q_PROPERTY accessors on Settings use Q_DECLARE_OPAQUE_POINTER so domain
headers don't need to be transitively included via settings.h.

Narrow consumers fully decoupled from settings.h:
- AutoWakeManager now takes SettingsAutoWake* (drops settings.h)
- DE1Device::setSettings() now takes SettingsHardware* (drops settings.h)
- AIManager listens to SettingsAI::configurationChanged

Settings.h: 1,170 → 887 lines (24% reduction). 4 remaining domains
(Brew, Dye, Network, App) deferred to a follow-up PR — see
openspec/changes/split-headers-by-domain/tasks.md for scope and rationale.

All 14 tests in tst_Settings pass; macOS build clean.
